+++ packages/client/src/encryption/GroupKeyManager.ts
@@ -1,6 +1,6 @@
 import { LitProtocolFacade } from './LitProtocolFacade'
 import { inject, Lifecycle, scoped } from 'tsyringe'
-import { GroupKeyStore } from './GroupKeyStore'
+import { LocalGroupKeyStore } from './LocalGroupKeyStore'
 import { GroupKey } from './GroupKey'
 import { StreamID, StreamPartID, StreamPartIDUtils } from '@streamr/protocol'
 import { EthereumAddress, waitForEvent } from '@streamr/utils'
@@ -10,29 +10,33 @@ import { StreamrClientEventEmitter } from '../events'
 import { DestroySignal } from '../DestroySignal'
 import crypto from 'crypto'
 import { uuid } from '../utils/uuid'
+import { Authentication, AuthenticationInjectionToken } from '../Authentication'
 
 @scoped(Lifecycle.ContainerScoped)
 export class GroupKeyManager {
-    private readonly groupKeyStore: GroupKeyStore
+    private readonly localGroupKeyStore: LocalGroupKeyStore
     private readonly litProtocolFacade: LitProtocolFacade
     private readonly subscriberKeyExchange: SubscriberKeyExchange
     private readonly eventEmitter: StreamrClientEventEmitter
     private readonly destroySignal: DestroySignal
+    private readonly authentication: Authentication
     private readonly config: Pick<StrictStreamrClientConfig, 'encryption'>
 
     constructor(
-        groupKeyStore: GroupKeyStore,
+        localGroupKeyStore: LocalGroupKeyStore,
         litProtocolFacade: LitProtocolFacade,
         subscriberKeyExchange: SubscriberKeyExchange,
         eventEmitter: StreamrClientEventEmitter,
         destroySignal: DestroySignal,
+        @inject(AuthenticationInjectionToken) authentication: Authentication,
         @inject(ConfigInjectionToken) config: Pick<StrictStreamrClientConfig, 'encryption'>
     ) {
-        this.groupKeyStore = groupKeyStore
+        this.localGroupKeyStore = localGroupKeyStore
         this.litProtocolFacade = litProtocolFacade
         this.subscriberKeyExchange = subscriberKeyExchange
         this.eventEmitter = eventEmitter
         this.destroySignal = destroySignal
+        this.authentication = authentication
         this.config = config
     }
 
@@ -40,7 +44,7 @@ export class GroupKeyManager {
         const streamId = StreamPartIDUtils.getStreamID(streamPartId)
 
         // 1st try: local storage
-        let groupKey = await this.groupKeyStore.get(groupKeyId, streamId)
+        let groupKey = await this.localGroupKeyStore.get(groupKeyId, publisherId)
         if (groupKey !== undefined) {
             return groupKey
         }
@@ -49,25 +53,37 @@ export class GroupKeyManager {
         if (this.config.encryption.litProtocolEnabled) {
             groupKey = await this.litProtocolFacade.get(streamId, groupKeyId)
             if (groupKey !== undefined) {
-                await this.groupKeyStore.add(groupKey, streamId)
+                await this.localGroupKeyStore.set(groupKey.id, publisherId, groupKey.data)
                 return groupKey
             }
         }
 
         // 3rd try: Streamr key-exchange
         await this.subscriberKeyExchange.requestGroupKey(groupKeyId, publisherId, streamPartId)
-        const groupKeys = await waitForEvent(
+        const groupKeyIds = await waitForEvent(
             // TODO remove "as any" type casing in NET-889
             this.eventEmitter as any,
-            'addGroupKey',
+            'storeEncryptionKeyToLocalStore',
             this.config.encryption.keyRequestTimeout,
-            (storedGroupKey: GroupKey) => storedGroupKey.id === groupKeyId,
+            (storedGroupKeyId: string) => storedGroupKeyId === groupKeyId,
             this.destroySignal.abortSignal
         )
-        return groupKeys[0] as GroupKey
+        groupKey = await this.localGroupKeyStore.get(groupKeyIds[0] as string, publisherId)
+        return groupKey!
     }
 
-    async storeKey(groupKey: GroupKey | undefined, streamId: StreamID): Promise<GroupKey> { // TODO: name
+    async fetchLatestEncryptionKey(publisherId: EthereumAddress, streamId: StreamID): Promise<GroupKey | undefined> {
+        if (publisherId !== (await this.authentication.getAddress())) {
+            throw new Error('storeKey: fetching latest encryption keys for other publishers not supported.')
+        }
+        const keyId = await this.localGroupKeyStore.getLatestEncryptionKeyId(publisherId, streamId)
+        return keyId !== undefined ? this.localGroupKeyStore.get(keyId, publisherId) : undefined
+    }
+
+    async storeKey(groupKey: GroupKey | undefined, publisherId: EthereumAddress, streamId: StreamID): Promise<GroupKey> {
+        if (publisherId !== (await this.authentication.getAddress())) { // TODO: unit test?
+            throw new Error('storeKey: storing latest encryption keys for other publishers not supported.')
+        }
         if (groupKey === undefined) {
             const keyData = crypto.randomBytes(32)
             // 1st try lit-protocol, if a key cannot be generated and stored, then generate group key locally
@@ -78,11 +94,12 @@ export class GroupKeyManager {
                 groupKey = new GroupKey(uuid('GroupKey'), keyData)
             }
         }
-        await this.groupKeyStore.add(groupKey, streamId)
+        await this.localGroupKeyStore.set(groupKey.id, publisherId, groupKey.data)
+        await this.localGroupKeyStore.setLatestEncryptionKeyId(groupKey.id, publisherId, streamId)
         return groupKey
     }
 
-    addKeyToLocalStore(groupKey: GroupKey, streamId: StreamID): Promise<void> {
-        return this.groupKeyStore.add(groupKey, streamId)
+    addKeyToLocalStore(groupKey: GroupKey, publisherId: EthereumAddress): Promise<void> {
+        return this.localGroupKeyStore.set(groupKey.id, publisherId, groupKey.data)
     }
 }