+++ packages/client/src/publish/GroupKeyQueue.ts
@@ -1,6 +1,7 @@
 import { StreamID } from '@streamr/protocol'
 import { GroupKey } from '../encryption/GroupKey'
 import { GroupKeyManager } from '../encryption/GroupKeyManager'
+import { Authentication } from '../Authentication'
 
 export interface GroupKeySequence {
     current: GroupKey
@@ -12,10 +13,25 @@ export class GroupKeyQueue {
     private currentGroupKey: GroupKey | undefined
     private queuedGroupKey: GroupKey | undefined // a group key queued to be rotated into use after the call to useGroupKey
     private readonly streamId: StreamID
+    private readonly authentication: Authentication
     private readonly groupKeyManager: GroupKeyManager
 
-    constructor(streamId: StreamID, groupKeyManager: GroupKeyManager) {
+    static async createInstance(
+        streamId: StreamID,
+        authentication: Authentication,
+        groupKeyManager: GroupKeyManager
+    ): Promise<GroupKeyQueue> {
+        const instance = new GroupKeyQueue(streamId, authentication, groupKeyManager)
+        instance.currentGroupKey = await instance.groupKeyManager.fetchLatestEncryptionKey(
+            await authentication.getAddress(),
+            streamId,
+        )
+        return instance
+    }
+
+    private constructor(streamId: StreamID, authentication: Authentication, groupKeyManager: GroupKeyManager) {
         this.streamId = streamId
+        this.authentication = authentication
         this.groupKeyManager = groupKeyManager
     }
 
@@ -39,13 +55,15 @@ export class GroupKeyQueue {
     }
 
     async rotate(newKey?: GroupKey): Promise<GroupKey> {
-        newKey = await this.groupKeyManager.storeKey(newKey, this.streamId)
+        const publisherId = await this.authentication.getAddress()
+        newKey = await this.groupKeyManager.storeKey(newKey, publisherId, this.streamId)
         this.queuedGroupKey = newKey
         return newKey
     }
 
     async rekey(newKey?: GroupKey): Promise<GroupKey> {
-        newKey = await this.groupKeyManager.storeKey(newKey, this.streamId)
+        const publisherId = await this.authentication.getAddress()
+        newKey = await this.groupKeyManager.storeKey(newKey, publisherId, this.streamId)
         this.currentGroupKey = newKey
         this.queuedGroupKey = undefined
         return newKey