+++ packages/client/src/utils/persistence/ServerPersistence.ts
@@ -6,27 +6,18 @@ import sqlite3 from 'sqlite3'
 
 import { pOnce } from '../promises'
 
-import { Persistence } from './Persistence'
-import { StreamID } from '@streamr/protocol'
+import { PersistenceContext, PersistenceContextOptions } from './PersistenceContext'
 import { Logger, wait } from '@streamr/utils'
 import { LoggerFactory } from '../LoggerFactory'
 
-export interface ServerPersistenceOptions {
+export interface ServerPersistenceOptions extends PersistenceContextOptions {
     loggerFactory: LoggerFactory
-    tableName: string
-    valueColumnName: string
-    clientId: string
     migrationsPath?: string
     onInit?: (db: Database) => Promise<void>
 }
 
-/*
- * Stores key-value pairs for a given stream
- */
-export default class ServerPersistence implements Persistence<string, string> {
+export default class ServerPersistence implements PersistenceContext {
     private readonly logger: Logger
-    private readonly tableName: string
-    private readonly valueColumnName: string
     private readonly dbFilePath: string
     private store?: Database
     private error?: Error
@@ -34,20 +25,25 @@ export default class ServerPersistence implements Persistence<string, string> {
     private readonly migrationsPath?: string
     private readonly onInit?: (db: Database) => Promise<void>
 
-    constructor({
+    // uses createInstance factory pattern so that ServerPersistence and BrowserPersistence
+    // are interchangeable
+    static async createInstance(opts: ServerPersistenceOptions): Promise<ServerPersistence> {
+        // TODO init() call could called here, so that we don't need to separate logic for 
+        // initialization (i.e. check this.initCalled flag before eaach call).
+        // It would be ok to do initialization, because the PersistenceManager already lazy loads
+        // and therefore doesn't create this instance before it is needed
+        return new ServerPersistence(opts)
+    }
+
+    private constructor({
         loggerFactory,
         clientId,
-        tableName,
-        valueColumnName,
         migrationsPath,
         onInit
     }: ServerPersistenceOptions) {
         this.logger = loggerFactory.createLogger(module)
-        this.tableName = tableName
-        this.valueColumnName = valueColumnName
         const paths = envPaths('streamr-client')
-        const dbFilePath = resolve(paths.data, join('./', clientId, `${tableName}.db`))
-        this.dbFilePath = dbFilePath
+        this.dbFilePath = resolve(paths.data, join('./', clientId, `GroupKeys.db`))
         this.migrationsPath = migrationsPath
         this.onInit = onInit
         this.init = pOnce(this.init.bind(this))
@@ -132,29 +128,27 @@ export default class ServerPersistence implements Persistence<string, string> {
         this.logger.trace('database initialized')
     }
 
-    async get(key: string, streamId: StreamID): Promise<string | undefined> {
+    async get(key: string, namespace: string): Promise<string | undefined> {
         if (!this.initCalled) {
             // can't have if doesn't exist
             if (!(await this.exists())) { return undefined }
         }
 
         await this.init()
-        const value = await this.store!.get(
-            `SELECT ${this.valueColumnName} FROM ${this.tableName} WHERE id = ? AND streamId = ?`,
-            key,
-            encodeURIComponent(streamId)
+        const row = await this.store!.get(
+            `SELECT value_ FROM ${namespace} WHERE key_ = ?`,
+            key
         )
-        return value?.[this.valueColumnName]
+        return row?.['value_']
     }
 
-    async set(key: string, value: string, streamId: StreamID): Promise<void> {
+    async set(key: string, value: string, namespace: string): Promise<void> {
         await this.init()
         await this.store!.run(
-            `INSERT INTO ${this.tableName} VALUES ($id, $${this.valueColumnName}, $streamId) ON CONFLICT DO NOTHING`,
+            `INSERT INTO ${namespace} (key_, value_) VALUES ($key_, $value_) ON CONFLICT DO UPDATE SET value_ = $value_`,
             {
-                $id: key,
-                [`$${this.valueColumnName}`]: value,
-                $streamId: encodeURIComponent(streamId),
+                $key_: key,
+                $value_: value,
             }
         )
     }
@@ -168,9 +162,5 @@ export default class ServerPersistence implements Persistence<string, string> {
         await this.init()
         await this.store!.close()
         this.logger.trace('closed')
-    }
-
-    get [Symbol.toStringTag](): string {
-        return this.constructor.name
     }
 }